Remove Picasso lib dependency from Buy SDK (#200)
